Move Storage into bob.storage pkg
