+++ src/main/java/bob/storage/Storage.java
@@ -1,4 +1,4 @@
-package bob;
+package bob.storage;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
