+++ src/main/java/bob/task/TaskList.java
@@ -2,7 +2,7 @@ package bob.task;
 import java.io.IOException;
 import java.util.List;
 
-import bob.Storage;
+import bob.storage.Storage;
 
 public class TaskList {
     private static List<Task> tasks;
